Implemented MCP4728 driver with passing pytests
This driver is implemented in Python. It uses the aiosmbus package
to talk to the device over i2c. An example of how to use this
module can be seen in test_mcp4728.py.
All tests are passing in this commit.
The driver does not implement device address read or update since
those functions require synchronous control of the LDAC line.
The General Call Software Update cannot be tested without feedback
of the DAC Vout.